Update type in rStringCopyUntil

+++ lib/rlib/string.c
@@ -5,7 +5,7 @@
 
 char *rStringCopyUntil(const char *from, char until) {
     const char *itEnd = strchr(from, until);
-    unsigned long lineLength = strnlen(from, itEnd - from);
+    size_t lineLength = strnlen(from, itEnd - from);
 
     char *line = calloc(lineLength + 1, sizeof(char));
     strlcpy(line, from, lineLength + 1);
